Temporarily changed to single precision to save new zker.

+++ Numerics_with_Fortran/write_2Darray.f90
@@ -109,7 +109,7 @@ END SUBROUTINE
 
 SUBROUTINE write_cmplx_2DArray_bin(data, filename)
     IMPLICIT NONE
-    COMPLEX(4), intent(in) :: data(:,:)
+    COMPLEX(8), intent(in) :: data(:,:)
     CHARACTER(len=*), intent(in) :: filename
 
     write(*,*) "Writing cx 2D to file ", filename
